Solved active mouse on entire page for the draw-spiral sketch.
With instance mode, I createCanvas with assignment to the 'canvas' variable.
I then overwrite the 'defaultCanvas' id with a unique 'canvas.id' of
'draw-spiral.' Then, pointing to the canvas element with its mousePressed
method, I provide the 'ds.newSpiral' callback function that now only reacts
when the mouse is clicked within the canvas DOM element.