Corrected the active mouse issue on fibonacci-static.
This change was tricky as the original sketch used the draw loop to
mimic a mouseDragged function. Oddly, though, canvas.mouseDragged was
returning in the browser console as 'not a function'. canvas.mousePressed
does not work as an effective call because in this way it only grabs
the mouse coordinates once, taking away the click, hold, and drag
drawing functionality. Using canvas.mouseMoved, the sketch now ignores
mouse clicks and draws a Fibonacci with mouse movement. This is corrected
by creating fst.mouseMoved function, which is called by the
fst.newFibonacci function only when fst.mouseIsPressed.