Created draw-spiral-2 sketch with fill, alpha, rotate, and scale --incomplete.
Have found in working with the scale factor, then additionally with the
frameRate and incRadius increment values, that the arc function is not
receiving polar-converted values for the minor and major axes of the partial
ellipse that an arc function builds. Will need to incorporate polar-to
-cartesian conversion for each draw loop to adjust the trajectory of the
arc segment to appear a contiguous spiral, not the sawtooth appearance
of current functionality.